Rename state to clarify naming
