+++ packages/twenty-front/src/modules/ui/layout/dropdown/states/internal/dropdownMaxHeightComponentState.ts
@@ -1,10 +1,10 @@
 import { DropdownComponentInstanceContext } from '@/ui/layout/dropdown/contexts/DropdownComponeInstanceContext';
 import { createComponentStateV2 } from '@/ui/utilities/state/component-state/utils/createComponentStateV2';
 
-export const dropdownMaxHeightComponentStateV2 = createComponentStateV2<
+export const dropdownMaxHeightComponentState = createComponentStateV2<
   number | undefined
 >({
-  key: 'dropdownMaxHeightComponentStateV2',
+  key: 'dropdownMaxHeightComponentState',
   componentInstanceContext: DropdownComponentInstanceContext,
   defaultValue: undefined,
 });
